More Entity updates, began the EntityFactor (still unfinished)

+++ src/main/java/net/canarymod/api/factory/CanaryBlockFactory.java
@@ -1,7 +1,6 @@
 package net.canarymod.api.factory;
 
 
-import net.canarymod.api.factory.BlockFactory;
 import net.canarymod.api.world.CanaryWorld;
 import net.canarymod.api.world.World;
 import net.canarymod.api.world.blocks.Block;
@@ -22,27 +21,27 @@ public class CanaryBlockFactory implements BlockFactory {
 
     @Override
     public Block newBlock(int id, World world) {
-        return new CanaryBlock((short) id, (byte) 0, 0, 0, 0, (CanaryWorld) world);
+        return new CanaryBlock((short) id, (short) 0, 0, 0, 0, (CanaryWorld) world);
     }
 
     @Override
     public Block newBlock(int id, int damage, World world) {
-        return new CanaryBlock((short) id, (byte) damage, 0, 0, 0, (CanaryWorld) world);
+        return new CanaryBlock((short) id, (short) damage, 0, 0, 0, (CanaryWorld) world);
     }
 
     @Override
     public Block newBlock(int id, int x, int y, int z) {
-        return new CanaryBlock((short) id, (byte) 0, x, y, z);
+        return new CanaryBlock((short) id, (short) 0, x, y, z);
     }
 
     @Override
     public Block newBlock(int id, int damage, int x, int y, int z) {
-        return new CanaryBlock((short) id, (byte) damage, x, y, z);
+        return new CanaryBlock((short) id, (short) damage, x, y, z);
     }
 
     @Override
     public Block newBlock(int id, int damage, int x, int y, int z, World world) {
-        return new CanaryBlock((short) id, (byte) damage, x, y, z, (CanaryWorld) world);
+        return new CanaryBlock((short) id, (short) damage, x, y, z, (CanaryWorld) world);
     }
 
     @Override
